api/requirements.txt
pyyaml>=6.0,<7.0
pandas>=3.0.1,<4.0.0
fpdf>=1.7.2,<2.0.0
openpyxl>=3.1.0,<4.0.0
fastapi>=0.115
uvicorn[standard]>=0.30
motor>=3.5
beanie>=1.27
python-dotenv>=1.0
pydantic-settings>=2.0
minio>=7.2.20,<8.0.0
anthropic>=0.84.0,<0.85.0
python-multipart>=0.0.22,<0.0.23
python-jose[cryptography]>=3.3,<4.0
bcrypt>=4.0
click>=8.0,<9.0
